manager/requirements.txt
# Core framework
py4web>=1.20240901.1
uvicorn[standard]==0.24.0
uvloop==0.19.0

# Async and HTTP
aiohttp==3.9.1
aiofiles==23.2.1
httpx==0.25.2

# Authentication and security
bcrypt>=4.1.2
cryptography==41.0.7
pyjwt>=2.8.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23

# Redis
redis==5.0.1
aioredis==2.0.1

# Metrics and monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Data handling
pydantic==2.5.3
pyyaml==6.0.1

# Logging
structlog==23.2.0

# Utilities
dnspython==2.4.2
python-dotenv==1.0.0

# Testing (dev dependencies)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Linting (dev dependencies)
pylint==3.0.3
mypy==1.7.1